chore: move resources form to separate component
