+++ ui/apps/everest/src/components/cluster-form/index.ts
@@ -1,0 +1,1 @@
+export { default as ResourcesForm } from './resources';
