+++ ui/apps/everest/src/components/cluster-form/resources/index.ts
@@ -1,0 +1,1 @@
+export { default } from './resources';
