+++ ui/apps/everest/src/pages/database-form/database-form-body/steps/first/first-step.tsx
@@ -32,7 +32,8 @@ import { useNamespaces } from 'hooks/api/namespaces/useNamespaces';
 import { useFormContext } from 'react-hook-form';
 import { DbEngineToolStatus } from 'shared-types/dbEngines.types';
 import { DB_WIZARD_DEFAULTS } from '../../../database-form.constants.ts';
-import { DbWizardFormFields, StepProps } from '../../../database-form.types.ts';
+import { StepProps } from '../../../database-form.types.ts';
+import { DbWizardFormFields } from 'consts.ts';
 import { useDatabasePageMode } from '../../../useDatabasePageMode.ts';
 import { StepHeader } from '../step-header/step-header.tsx';
 import { DEFAULT_NODES } from './first-step.constants.ts';
