--- ui/apps/everest/src/pages/database-form/database-form-body/steps/resources/resource-input/resource-input.types.ts
@@ -1,8 +1,0 @@
-export type ResourceInputProps = {
-  name: string;
-  label: string;
-  helperText: string;
-  endSuffix: string;
-  numberOfNodes: number;
-  disabled?: boolean;
-};
