--- ui/apps/everest/src/pages/database-form/database-form-body/steps/resources/resources-step.utils.test.ts
@@ -1,25 +1,0 @@
-// everest
-// Copyright (C) 2023 Percona LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-import { getResourceNames } from './resources-step.utils.ts';
-
-describe('Resources step utils', () => {
-  it('should show correct alert message', async () => {
-    expect(getResourceNames(['CPU', 'memory', 'disk'])).toBe(
-      'CPU, memory, and disk'
-    );
-    expect(getResourceNames(['memory', 'disk'])).toBe('memory and disk');
-    expect(getResourceNames(['disk'])).toBe('disk');
-  });
-});
